refactor: create new files for nodes and edges setting
